cli: refactor graph submod to share meta code with log
