+++ modality-probe-cli/src/export/mod.rs
@@ -1,22 +1,16 @@
 //! Export a textual representation of a causal graph using the
 //! collected columnar form as input.
 
-use std::{collections::HashMap, fmt, fs::File, path::PathBuf, str::FromStr};
+use std::{fs::File, path::PathBuf, str::FromStr};
 
-use err_derive::Error;
 use structopt::StructOpt;
-use uuid::Uuid;
 
-use modality_probe::ProbeId;
-use modality_probe_collector_common::{json, Error as CollectorError, LogEntryData};
-use modality_probe_graph::GraphEvent;
+use modality_probe_collector_common::json;
 
-use crate::{component::Component, events::Events};
+use crate::{give_up, hopefully, meta};
 
+pub mod graph;
 mod templates;
-
-mod graph;
-use graph::{EventMeta, ProbeMeta};
 
 /// Export a textual representation of a causal graph using the
 /// collected trace file as input.
@@ -46,21 +40,6 @@ pub struct Export {
     pub graph_type: GraphType,
 }
 
-/// Inspect the event log from the perspective of a single probe.
-#[derive(Debug, PartialEq, StructOpt)]
-pub struct Log {
-    // /// The probe to target.
-    // #[structopt(short, long)]
-    // pub probe: String,
-    /// The path to a component directory. To include multiple
-    /// components, provide this switch multiple times.
-    #[structopt(short, long, required = true)]
-    pub components: Vec<PathBuf>,
-    /// The path to the collected trace.
-    #[structopt(short, long, required = true)]
-    pub report: PathBuf,
-}
-
 #[derive(Debug, PartialEq, StructOpt)]
 pub enum GraphType {
     Cyclic,
@@ -68,209 +47,22 @@ pub enum GraphType {
 }
 
 impl FromStr for GraphType {
-    type Err = ExportError;
-    fn from_str(s: &str) -> Result<Self, ExportError> {
+    type Err = Box<dyn std::error::Error>;
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
         match s {
             "cyclic" => Ok(GraphType::Cyclic),
             "acyclic" => Ok(GraphType::Acyclic),
-            _ => Err(ExportError(format!("{} is not a valid graph type", s))),
+            _ => give_up!(format!("{} is not a valid graph type", s)),
         }
     }
 }
 
-#[derive(Debug, Error)]
-#[error(display = "{}", _0)]
-pub struct ExportError(String);
-
-impl From<fmt::Error> for ExportError {
-    fn from(e: fmt::Error) -> Self {
-        ExportError(format!("encountered an error generating dot: {}", e))
-    }
-}
-
-impl From<CollectorError> for ExportError {
-    fn from(e: CollectorError) -> Self {
-        ExportError(format!("failed to read log file: {}", e))
-    }
-}
-
-pub fn log_all(mut l: Log) -> Result<(), ExportError> {
-    let cfg = assemble_components(&mut l.components)?;
-    let mut log_file = File::open(&l.report).map_err(|e| {
-        ExportError(format!(
-            "failed to open the report file at {}: {}",
-            l.report.display(),
-            e
-        ))
-    })?;
-
-    let report = json::read_log_entries(&mut log_file)?;
-    let mut probes = HashMap::new();
-    for ev in report {
-        let p = probes.entry(ev.probe_id).or_insert_with(Vec::new);
-        p.push(ev);
-    }
-
-    // Sort the probe-sorted events by sequence.
-    for plog in probes.values_mut() {
-        plog.sort_by_key(|p| std::cmp::Reverse((p.sequence_number, p.sequence_index)));
-    }
-
-    let mut log = Vec::new();
-    // Round-robin through the probe-sorted list and add events to the
-    // log, interleaved probe by probe, but maintaining their total
-    // order.
-    loop {
-        let init_len = log.len();
-        for plog in probes.values_mut() {
-            if let Some(ev) = plog.pop() {
-                log.push(ev);
-            }
-        }
-        if log.len() == init_len {
-            break;
-        }
-    }
-    // Print the log.
-    for l in log {
-        match l.data {
-            LogEntryData::Event(id) => {
-                let emeta = graph::get_event_meta(&cfg, &l.probe_id, &id)?;
-                let probe_name = cfg
-                    .probes
-                    .get(&l.probe_id.get_raw())
-                    .map(|p| p.name.clone())
-                    .unwrap_or(format!("UNKNOWN_PROBE_{}", l.probe_id.get_raw()));
-                print_log_line(emeta, &probe_name, None);
-            }
-            LogEntryData::EventWithPayload(id, pl) => {
-                let emeta = graph::get_event_meta(&cfg, &l.probe_id, &id)?;
-                let probe_name = cfg
-                    .probes
-                    .get(&l.probe_id.get_raw())
-                    .map(|p| p.name.clone())
-                    .unwrap_or(format!("UNKNOWN_PROBE_{}", l.probe_id.get_raw()));
-                print_log_line(
-                    emeta,
-                    &probe_name,
-                    Some(graph::parsed_payload(
-                        &emeta.type_hint.clone().unwrap_or_else(|| pl.to_string()),
-                        pl,
-                    )?),
-                );
-            }
-            _ => (),
-        }
-    }
-    Ok(())
-}
-
-fn print_log_line(emeta: &EventMeta, probe_name: &str, payload: Option<String>) {
-    println!(
-        "{} probe={} description={}{}{}tags={}{}",
-        emeta.name,
-        probe_name,
-        emeta.description,
-        if !emeta.file.is_empty() {
-            format!(" file={} ", emeta.file)
-        } else {
-            "".to_string()
-        },
-        if !emeta.line.is_empty() {
-            format!(" line={} ", emeta.line)
-        } else {
-            "".to_string()
-        },
-        emeta.tags,
-        match payload {
-            Some(p) => format!(" payload={}", p),
-            None => "".to_string(),
-        }
-    );
-}
-
-// pub fn log(mut l: Log) -> Result<(), ExportError> {
-//     let cfg = assemble_components(&mut l.components)?;
-//     let mut log_file = File::open(&l.report).map_err(|e| {
-//         ExportError(format!(
-//             "failed to open the report file at {}: {}",
-//             l.report.display(),
-//             e
-//         ))
-//     })?;
-
-//     let graph = graph::log_to_graph(
-//         json::read_log_entries(&mut log_file)?
-//             .into_iter()
-//             .peekable(),
-//     )?;
-
-//     let probe = match cfg.probes.iter().find(|(_, v)| v.name == l.probe) {
-//         Some((_, pm)) => pm,
-//         None => {
-//             let pid = l
-//                 .probe
-//                 .parse::<u32>()
-//                 .map_err(|_| ExportError(format!("probe {} could not be found", l.probe)))?;
-//             cfg.probes
-//                 .get(&pid)
-//                 .ok_or_else(|| ExportError(format!("probe {} could not be found", l.probe)))?
-//         }
-//     };
-//     let log = graph.graph.probe_log(ProbeId::new(probe.id).ok_or_else(|| {
-//         ExportError(format!(
-//             "encountered the invalid probe id {} when looking up probe {}",
-//             probe.id, l.probe
-//         ))
-//     })?);
-//     print_log(&log, &cfg)
-// }
-
-// fn print_log(log: &Vec<&GraphEvent>, cfg: &Cfg) -> Result<(), ExportError> {
-//     for l in log {
-//         let emeta = graph::get_event_meta(cfg, &l.probe_id, &l.id)?;
-//         let probe_name = cfg
-//             .probes
-//             .get(&l.probe_id.get_raw())
-//             .map(|p| p.name.clone())
-//             .unwrap_or(format!("UNKNOWN_PROBE_{}", l.probe_id.get_raw()));
-//         println!(
-//             "{} probe={} description={}{}{}tags={}{}",
-//             emeta.name,
-//             probe_name,
-//             emeta.description,
-//             if !emeta.file.is_empty() {
-//                 format!(" file={} ", emeta.file)
-//             } else {
-//                 "".to_string()
-//             },
-//             if !emeta.line.is_empty() {
-//                 format!(" line={} ", emeta.line)
-//             } else {
-//                 "".to_string()
-//             },
-//             emeta.tags,
-//             match l.payload {
-//                 Some(p) => format!(
-//                     " payload={}",
-//                     graph::parsed_payload(&emeta.type_hint.as_ref().unwrap(), p)?
-//                 ),
-//                 None => "".to_string(),
-//             }
-//         );
-//     }
-//     Ok(())
-// }
-
-pub fn run(mut exp: Export) -> Result<(), ExportError> {
-    let cfg = assemble_components(&mut exp.components)?;
-    let mut log_file = File::open(&exp.report).map_err(|e| {
-        ExportError(format!(
-            "failed to open the report file at {}: {}",
-            exp.report.display(),
-            e
-        ))
-    })?;
+pub fn run(mut exp: Export) -> Result<(), Box<dyn std::error::Error>> {
+    let cfg = meta::assemble_components(&mut exp.components)?;
+    let mut log_file = hopefully!(
+        File::open(&exp.report),
+        format!("failed to open the report file at {}", exp.report.display(),)
+    )?;
 
     let graph = graph::log_to_graph(
         json::read_log_entries(&mut log_file)?
@@ -312,243 +104,4 @@ pub fn run(mut exp: Export) -> Result<(), ExportError> {
     }
 
     Ok(())
-}
-
-pub struct Cfg {
-    pub probes: HashMap<u32, ProbeMeta>,
-    pub events: HashMap<(Uuid, u32), EventMeta>,
-    pub probes_to_components: HashMap<u32, Uuid>,
-    pub component_names: HashMap<String, String>,
-}
-
-fn assemble_components(comp_dirs: &mut Vec<PathBuf>) -> Result<Cfg, ExportError> {
-    let mut probes = HashMap::new();
-    let mut probes_to_components = HashMap::new();
-    let mut events = HashMap::new();
-    let mut component_names = HashMap::new();
-
-    let mut event_files = Vec::new();
-    let mut probe_files = Vec::new();
-    for dir in comp_dirs.iter_mut() {
-        dir.push("probes.csv");
-        probe_files.push(dir.clone());
-        dir.pop();
-        dir.push("events.csv");
-        event_files.push(dir.clone());
-    }
-    for pf in probe_files.iter_mut() {
-        let mut pr_rdr = csv::Reader::from_path(pf.clone()).map_err(|e| {
-            ExportError(format!(
-                "failed to open the probes file at {}: {}",
-                pf.display(),
-                e
-            ))
-        })?;
-        for res in pr_rdr.deserialize() {
-            let p: ProbeMeta =
-                res.map_err(|e| ExportError(format!("failed to deserialize a probe row: {}", e)))?;
-            pf.pop();
-            pf.push("Component.toml");
-            let comp = Component::from_toml(&pf);
-            probes.insert(p.id, p.clone());
-            probes_to_components.insert(p.id, comp.id.0);
-            component_names.insert(comp.id.0.to_string(), comp.name);
-        }
-    }
-    for ef in event_files.iter_mut() {
-        let mut ev_rdr = csv::Reader::from_path(ef.clone()).map_err(|e| {
-            ExportError(format!(
-                "failed to open the events file at {}: {}",
-                ef.display(),
-                e
-            ))
-        })?;
-        for res in ev_rdr.deserialize() {
-            let e: EventMeta =
-                res.map_err(|e| ExportError(format!("failed to deserialize a event row: {}", e)))?;
-            ef.pop();
-            ef.push("Component.toml");
-            let comp = Component::from_toml(&ef);
-            events.insert((comp.id.0, e.id), e.clone());
-        }
-    }
-
-    add_internal_events(&mut events)?;
-
-    Ok(Cfg {
-        events,
-        probes,
-        probes_to_components,
-        component_names,
-    })
-}
-
-fn add_internal_events(events: &mut HashMap<(Uuid, u32), EventMeta>) -> Result<(), ExportError> {
-    let nil_uuid = Uuid::nil();
-    for ie in Events::internal_events() {
-        let ev = EventMeta {
-            component_id: nil_uuid,
-            id: ie.id.0,
-            name: ie.name,
-            type_hint: if ie.type_hint.is_empty() {
-                None
-            } else {
-                Some(ie.type_hint)
-            },
-            tags: ie.tags,
-            description: ie.description,
-            file: ie.file,
-            line: ie.line,
-        };
-        events.insert((nil_uuid, ie.id.0), ev.clone());
-    }
-
-    Ok(())
-}
-
-#[cfg(test)]
-mod test {
-    use std::{fs, fs::File, io::Write};
-
-    use tempfile::tempdir;
-    use uuid::Uuid;
-
-    use super::graph::{EventMeta, ProbeMeta};
-
-    const COMP_ONE_CONTENT: &'static str = r#"
-name = "one"
-id = "bba61171-e4b5-4db4-8cbb-8b4f4a581ca1"
-"#;
-    const COMP_TWO_CONTENT: &'static str = r#"
-name = "two"
-id = "bba61171-e4b5-4db4-8cbb-8b4f4a581ca2"
-"#;
-    const PROBE_ONE_CONTENT: &'static str = "
-component_id,id,name,description,tags,file,line
-bba61171-e4b5-4db4-8cbb-8b4f4a581cb1,1,PROBE_ONE,probe one,example,examples/event-recording.rs,26";
-    const PROBE_TWO_CONTENT: &'static str = "
-component_id,id,name,description,tags,file,line
-bba61171-e4b5-4db4-8cbb-8b4f4a581cb2,2,PROBE_TWO,probe two,example,examples/event-recording.rs,26";
-
-    const EVENT_ONE_CONTENT: &'static str = "
-component_id,id,name,description,tags,type_hint,file,line
-bba61171-e4b5-4db4-8cbb-8b4f4a581cb1,1,TEST_ONE,test event one,,,,26
-";
-    const EVENT_TWO_CONTENT: &'static str = "
-component_id,id,name,description,tags,type_hint,file,line
-bba61171-e4b5-4db4-8cbb-8b4f4a581cb2,2,TEST_TWO,test event two,,,,36
-";
-
-    #[test]
-    fn comp_assembly() {
-        let mut tmp_one = tempdir().unwrap().into_path();
-        let mut tmp_two = tempdir().unwrap().into_path();
-
-        {
-            tmp_one.push("Component.toml");
-            let mut comp_one = File::create(&tmp_one).unwrap();
-            write!(comp_one, "{}", COMP_ONE_CONTENT).unwrap();
-            tmp_one.pop();
-            tmp_one.push("probes.csv");
-            let mut probe_one = File::create(&tmp_one).unwrap();
-            write!(probe_one, "{}", PROBE_ONE_CONTENT).unwrap();
-            tmp_one.pop();
-            tmp_one.push("events.csv");
-            let mut event_one = File::create(&tmp_one).unwrap();
-            write!(event_one, "{}", EVENT_ONE_CONTENT).unwrap();
-
-            tmp_two.push("Component.toml");
-            let mut comp_two = File::create(&tmp_two).unwrap();
-            write!(comp_two, "{}", COMP_TWO_CONTENT).unwrap();
-            tmp_two.pop();
-            tmp_two.push("probes.csv");
-            let mut probe_two = File::create(&tmp_two).unwrap();
-            write!(probe_two, "{}", PROBE_TWO_CONTENT).unwrap();
-            tmp_two.pop();
-            tmp_two.push("events.csv");
-            let mut event_two = File::create(&tmp_two).unwrap();
-            write!(event_two, "{}", EVENT_TWO_CONTENT).unwrap();
-
-            let expected_probes = vec![
-                (
-                    1,
-                    ProbeMeta {
-                        component_id: Uuid::parse_str("bba61171-e4b5-4db4-8cbb-8b4f4a581cb1")
-                            .unwrap(),
-                        id: 1,
-                        name: "PROBE_ONE".to_string(),
-                        description: "probe one".to_string(),
-                        file: "examples/event-recording.rs".to_string(),
-                        line: "26".to_string(),
-                    },
-                ),
-                (
-                    2,
-                    ProbeMeta {
-                        component_id: Uuid::parse_str("bba61171-e4b5-4db4-8cbb-8b4f4a581cb2")
-                            .unwrap(),
-                        id: 2,
-                        name: "PROBE_TWO".to_string(),
-                        description: "probe two".to_string(),
-                        file: "examples/event-recording.rs".to_string(),
-                        line: "26".to_string(),
-                    },
-                ),
-            ]
-            .into_iter()
-            .collect();
-
-            let mut expected_events = vec![
-                (
-                    (
-                        Uuid::parse_str("bba61171-e4b5-4db4-8cbb-8b4f4a581ca1").unwrap(),
-                        1,
-                    ),
-                    EventMeta {
-                        component_id: Uuid::parse_str("bba61171-e4b5-4db4-8cbb-8b4f4a581cb1")
-                            .unwrap(),
-                        id: 1,
-                        name: "TEST_ONE".to_string(),
-                        description: "test event one".to_string(),
-                        type_hint: None,
-                        tags: String::new(),
-                        file: String::new(),
-                        line: "26".to_string(),
-                    },
-                ),
-                (
-                    (
-                        Uuid::parse_str("bba61171-e4b5-4db4-8cbb-8b4f4a581ca2").unwrap(),
-                        2,
-                    ),
-                    EventMeta {
-                        component_id: Uuid::parse_str("bba61171-e4b5-4db4-8cbb-8b4f4a581cb2")
-                            .unwrap(),
-                        id: 2,
-                        name: "TEST_TWO".to_string(),
-                        description: "test event two".to_string(),
-                        type_hint: None,
-                        tags: String::new(),
-                        file: String::new(),
-                        line: "36".to_string(),
-                    },
-                ),
-            ]
-            .into_iter()
-            .collect();
-
-            super::add_internal_events(&mut expected_events).unwrap();
-
-            tmp_one.pop();
-            tmp_two.pop();
-            let cfg =
-                super::assemble_components(&mut vec![tmp_one.clone(), tmp_two.clone()]).unwrap();
-
-            assert_eq!(&cfg.probes, &expected_probes);
-            assert_eq!(&cfg.events, &expected_events);
-        }
-
-        fs::remove_dir_all(&tmp_one).unwrap();
-        fs::remove_dir_all(&tmp_two).unwrap();
-    }
 }
