+++ modality-probe-cli/src/main.rs
@@ -1,4 +1,4 @@
-use modality_probe_cli::{error::GracefulExit, export, header_gen, manifest_gen, opts::Opts};
+use modality_probe_cli::{error::GracefulExit, export, header_gen, log, manifest_gen, opts::Opts};
 use structopt::StructOpt;
 
 fn main() {
@@ -8,6 +8,6 @@ fn main() {
         Opts::ManifestGen(opt) => manifest_gen::run(opt, None),
         Opts::HeaderGen(opt) => header_gen::run(opt, None),
         Opts::Export(opt) => export::run(opt).unwrap_or_exit("export"),
-        Opts::Log(opt) => export::log_all(opt).unwrap_or_exit("export"),
+        Opts::Log(opt) => log::log_all(opt).unwrap_or_exit("export"),
     }
 }
